+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.1.0 (2025-02-11)
+
+- Add `easeInBack`, `easeOutBack` and `easeInOutBack` functions
+- Add `easeInBounce`, `easeOutBounce` and `easeInOutBounce` functions
+- Add `easeInCirc`, `easeOutCirc` and `easeInOutCirc` functions
+- Add `easeInElastic`, `easeOutElastic` and `easeInOutElastic` functions
+- Add `easeInExpo`, `easeOutExpo` and `easeInOutExpo` functions
+- Add `easeInQuint`, `easeOutQuint` and `easeInOutQuint` functions
+- Add `easeInSine`, `easeOutSine` and `easeInOutSine` functions
+
 ## 1.0.0 (2025-01-19)
 
-- First release of the package
+- Add `easeInCubic`, `easeOutCubic` and `easeInOutCubic` functions
+- Add `easeInQuad`, `easeOutQuad` and `easeInOutQuad` functions
+- Add `easeInQuart`, `easeOutQuart` and `easeInOutQuart` functions